refactor: moved DataPagination and DataTable components from agents to global components
